// apps/client/src/game/terminus-swarm/types.ts
export interface Vec2 {
  x: number;
  y: number;
}
// ...
export interface GameResources {
  salvage: number;
  viralIchor: number;
  neuralCores: number;
  voidCrystals: number;
}